pitext_travel/requirements.txt
# Core Flask dependencies
Flask==2.3.2
requests==2.31.0
openai==1.12.0
python-dotenv==1.0.0

# WebSocket and real-time communication - FIXED VERSIONS with ASGI support
flask-socketio==5.5.1
python-engineio==4.11.0
python-socketio>=5.12.0

# Audio processing
numpy==1.24.3

# CORS support
flask-cors==4.0.0

# Reliability
tenacity==8.2.3

# ASGI support
asgiref==3.7.2